First version of the reminder app
Using the Serverless framework, the app publishes to an AWS
Lambda that is scheduled to run every hour. Each hour the Tesla
API is check to see if the car is close to home and not in either
a Charging state or a Complete state. If the car is close to home
and in neither of those states, an email is sent to remind the user
to plug the car in.